+++ src/main/java/lotto/dto/StatisticDto.java
@@ -3,5 +3,5 @@ package lotto.dto;
 import java.util.EnumMap;
 import lotto.domain.rule.Rank;
 
-public record StatisticDto(EnumMap<Rank, Integer> statisticSummary, float profitRate) {
+public record StatisticDto(EnumMap<Rank, Integer> statisticSummary, double profitRate) {
 }
